Finally figured out how to auto-create the schema on a MySQL8 DigitalOcean server.
If we ever decide to fully migrate to digital ocean due to the unreliability of the TUDelft MySQL server.

+++ transactions/src/main/java/nl/tudelft/sem/transactions/controllers/TransactionController.java
@@ -83,7 +83,7 @@ public class TransactionController {
      */
     @DeleteMapping("/deleteTransaction/{transactionId}")
     @ResponseBody
-    public void deleteTransaction(@PathVariable(value = "transactionId") int transactionId) {
+    public void deleteTransaction(@PathVariable(value = "transactionId") long transactionId) {
         try {
             Optional<Transactions> t = transactionsRepository.findById(transactionId);
             Transactions transaction = t.get();
